Revert "Add tests for deserializing octet string xml with various types of whitespace"
This reverts commit 28a2dbcdfebf73d441223e5109415415b924ea59.
